pyproject.toml
[project]
name = "template-python"
version = "0.0.1"
description = "A GitHub template repository for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "typer>=0.16.0",
]

[dependency-groups]
docs = [
    "mkdocs-material>=9.6.14",
    "mkdocs-static-i18n>=1.3.0",
]
dev = [
    "coverage>=7.8.0",
    "jupyterlab>=4.4.2",
    "pre-commit>=4.2.0",
    "pyrefly>=0.44.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.7",
    "ty>=0.0.1a6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["D203"]

[tool.pytest.ini_options]
addopts = "-ra --cov"
testpaths = ["tests"]
pythonpath = ['.']
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.report]
show_missing = true

[tool.ty]
environment = { python-version = "3.10" }

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]
